mesto-one-13p-1

+++ controllers/cards.js
@@ -1,38 +1,38 @@
+const mongoose = require('mongoose');
 const card = require('../models/card');
 
 const getCard = (req, res) => {
   card
     .find({})
     .populate('owner')
-    .then((card) => {
-      res.send({ data: card });
-    })
+    .then((card) => res.send({ data: card }))
     .catch(() => res.status(500).send({ message: 'Server error' }));
 };
 
 const getDeleteCard = (res, req) => {
-  card
+  if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
+    return res.status(400).send({ message: 'Id is not correct' });
+  }
+  return card
     .findByIdAndRemove(req.params.cardId)
     .then((card) => {
       if (!card) {
         return res.status(404).send({ message: 'Card not found' });
       }
-      res.send({ data: card });
+      return res.send({ data: card });
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: "Id isn't correct" });
-      }
-      return res.status(500).send({ message: 'Server error' });
-    });
+    .catch(() => res.status(500).send({ message: 'Server error' }));
 };
 
 const createCard = (res, req) => {
   const { name, link } = req.body;
-
-  card
-    .create({ name, link, owner: req.user._id })
-    .then((card) => res.status(201).send(card))
+  const owner = req.user._id;
+  if (!name || !link) {
+    return res.status(400).send({ message: 'Error' });
+  }
+  return card
+    .create({ name, link, owner })
+    .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
         return res.status(400).send({ message: "Id isn't correct" });
@@ -42,7 +42,10 @@ const createCard = (res, req) => {
 };
 
 const likeCard = (req, res) => {
-  card
+  if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
+    return res.status(400).send({ message: 'Id is not correct' });
+  }
+  return card
     .findByIdAndUpdate(
       req.params.cardId,
       { $addToSet: { likes: req.user._id } },
@@ -52,33 +55,30 @@ const likeCard = (req, res) => {
       if (!card) {
         return res.status(404).send({ message: 'User not found' });
       }
-      return res.status(200).send({ data: card });
+      return res.send({ data: card });
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: "Id isn't correct" });
-      }
+    .catch(() => {
       return res.status(500).send({ message: 'Server error' });
     });
 };
 
 const disLikeCard = (req, res) => {
-  card
+  if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
+    return res.status(400).send({ message: 'Id is not correct' });
+  }
+  return card
     .findByIdAndUpdate(
       req.params.cardId,
       { $pull: { likes: req.user._id } },
-      { new: true },
+      {},
     )
     .then((card) => {
       if (!card) {
         return res.status(404).send({ message: ' User not found' });
       }
-      return res.status(200).send({ data: card });
+      return res.send({ data: card });
     })
-    .catch((err) => {
-      if (err.kind === 'CastError') {
-        return res.status(400).send({ message: "Id isn't correct" });
-      }
+    .catch(() => {
       return res.status(500).send({ message: 'Server error' });
     });
 };
